The correct approach for the 10 year sample size.
Originally, the goal was to see if we could use deep learning/neural networks in PyTorch. However, after a bit of trail and error, it was determined that regression and simple machine learning regression model would work much better for the amount of data we had (10 years of monthly data). This model accounts for the lag time between variables, uses multi-variate linear regression, and curve fitting to make sure the model is accurate.