Slight update to M1 data and testing of model
It was noticed when testing the model that:
A) Either the St. Louis fed had changed the format of the numbering
B) There was some inaccurate division of the numbers that were coming in
So this uploaded a new dataset to ensure that the numbers were consistent with current pulls of the M1 data.
This change also integrated more massaging of the data to ensure indexing wasn't listed as a column here. The model specifically needs 3 columns of input data and the massaging in the bottom cell accomplishes that.